bourses_criteres_sociaux/requirements.txt
pyexcel-ods==0.6.0
pyyaml==5.4.1
